requirements.txt
# SEA Kindergarten simulation (Phase 1)
mujoco>=3.0.0
numpy>=1.20.0
pyyaml>=6.0
